carve off shared code to helper functions
